requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
#streamlit==1.28.1
streamlit==1.18.1
openai==1.3.7
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0
pydantic==2.4.2
requests==2.31.0
pandas==2.0.3